% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Abbrev}
\alias{Abbrev}
\title{A data frame with metabolite naming conventions used in this package.}
\format{A data frame with seven variables and 204 entries}
\usage{
Abbrev
}
\description{
A data frame with metabolite naming conventions used in this package.
}
\keyword{datasets}
