% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBapps.R
\name{isomer}
\alias{isomer}
\title{This function takes in a chemical formula and will return all possible isomers based on the
KEGG database. Future versions might include the possibility to call the function
with a name or KEGG ID, too.}
\usage{
isomer(formula)
}
\arguments{
\item{formula}{is the chemical formula entered with quotation marks, ie as string}
}
\value{
a data frame of all potential isomers of that particular chemical formula
}
\description{
This function takes in a chemical formula and will return all possible isomers based on the
KEGG database. Future versions might include the possibility to call the function
with a name or KEGG ID, too.
}
\details{
Whatever details I might want to add...
}
\author{
Daniel Braas
}
\seealso{
\code{\link{keggFind}}
}

