% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_FC.R
\name{make_unlabeled}
\alias{make_unlabeled}
\title{This function produces a data frame for unlabeled data.}
\usage{
make_unlabeled(DF)
}
\arguments{
\item{DF}{The input data. This should be a data frame that contains MID data.}
}
\value{
A data frame with fractional contribution data.
}
\description{
This function produces a data frame with fractional contribution data.
}
\author{
Daniel Braas
}

