% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBapps.R
\name{replace_character}
\alias{replace_character}
\title{This function takes in a string and either replaces or inserts a character string.}
\usage{
replace_character(string, position, replacement)
}
\arguments{
\item{string}{The original string to use.}

\item{position}{The position at which to replace or insert the characters.}

\item{replacement}{The character string the will be inserted.}
}
\value{
A vector (within a data table) with re-mapped values
}
\description{
This function takes in a string and either replaces or inserts a character string.
}
